refactor: migrate embeddings implementation into drift-engine
(cherry picked from commit 52bab5b898f169ea9e35f88df85d7bfb9edcb005)
